AMP-22195: add null value workarounds
git-svn-id: http://svn.dgfoundation.org/amp/branches/AMP_2_11_RELEASE@29524 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/dgfoundation/amp/currencyconvertor/AmpCurrencyConvertor.java
@@ -90,7 +90,7 @@ public class AmpCurrencyConvertor implements CurrencyConvertor {
 			return getExchangeRate(baseCurrencyCode, toCurrencyCode, null, date) / fixedExchangeRate; 
 		}
 		
-		long julianCode = DateTimeUtil.toJulianDayNumber(date);
+		Integer julianCode = DateTimeUtil.toJulianDayNumber(date);
 		double res = getCalculator(fromCurrencyCode).getRate(julianCode) / getCalculator(toCurrencyCode).getRate(julianCode);
 		return res;
 	}
